--- lib/discourse_data_explorer/engine.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module DiscourseDataExplorer
-  class Engine < ::Rails::Engine
-    isolate_namespace DiscourseDataExplorer
-  end
-end
