+++ lib/tasks/data_explorer.rake
@@ -2,7 +2,6 @@
 
 # rake data_explorer:list_hidden_queries
 desc "Shows a list of hidden queries"
-task("data_explorer:list_hidden_queries").clear
 task "data_explorer:list_hidden_queries" => :environment do |t|
   puts "\nHidden Queries\n\n"
 
@@ -18,7 +17,6 @@ end
 # rake data_explorer[-1]
 # rake data_explorer[1,-2,3,-4,5]
 desc "Hides one or multiple queries by ID"
-task("data_explorer").clear
 task "data_explorer" => :environment do |t, args|
   args.extras.each do |arg|
     id = arg.to_i
@@ -37,7 +35,6 @@ end
 # rake data_explorer:unhide_query[-1]
 # rake data_explorer:unhide_query[1,-2,3,-4,5]
 desc "Unhides one or multiple queries by ID"
-task("data_explorer:unhide_query").clear
 task "data_explorer:unhide_query" => :environment do |t, args|
   args.extras.each do |arg|
     id = arg.to_i
@@ -56,7 +53,6 @@ end
 # rake data_explorer:hard_delete[-1]
 # rake data_explorer:hard_delete[1,-2,3,-4,5]
 desc "Hard deletes one or multiple queries by ID"
-task("data_explorer:hard_delete").clear
 task "data_explorer:hard_delete" => :environment do |t, args|
   args.extras.each do |arg|
     id = arg.to_i
